Add comments for all code and modules
This adds comments and some helpful remarks to all of the code that is
worthwhile.  Parts of the `RandomCode` module need to be deleted and
were not commented.  I am using MZ Tools to generate the comment blocks.
There is an extra item called `Flag` which indicates some additional
info for Subs that are not-used or are new-feature and not added to the
Ribbon yet.  I am also using an indent helper add-in to make all the
indentation consistent.  See Office Automation Smart Indenter.
Addresses #16.